Refactor local metadata implementation
